+++ src/test/java/com/historycode/ui/adminPanel/NewsPage/CreateNewsWithoutMandatoryFields.java
@@ -1,4 +1,4 @@
-package com.historycode.ui.adminPanel;
+package com.historycode.ui.adminPanel.NewsPage;
 
 import com.historycode.ui.page.adminpanel.newspage.NewsPageAdminPanel;
 import com.historycode.ui.page.adminpanel.newspage.NewsPageGridComponent;
@@ -16,55 +16,62 @@ import java.util.Random;
 
 import static org.testng.Assert.*;
 
-public class CreateNewsTest extends BaseTestRunnerWithAdmin {
+public class CreateNewsWithoutMandatoryFields extends BaseTestRunnerWithAdmin {
 
     private String createdTitle;
     private String createdLink;
     private String createdText;
+    private Date createdDate;
     private String imagePath = "src/test/resources/newsTest.png";
 
     @BeforeMethod
     public void beforeMethod() {
-        
-
         Random rand = new Random();
         int n = rand.nextInt(50);
         createdTitle = "Тестова новина " + n;
         createdLink = "test-link-" + n;
-        createdText = "Lorem ipsum dolor sit amet, consectetur adipiscing elit.";
+        createdText = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras et commodo ex. Pellentesque id sagittis ex. Morbi tincidunt volutpat ante, ut elementum turpis pulvinar et.";
+        createdDate = new Date(System.currentTimeMillis());
 
         driver.get(testValueProvider.getBaseUIUrl() + "/admin-panel/news");
-        NewsPageAdminPanel newsPage = new NewsPageAdminPanel(driver);
-
-        CreateEditNewsModal editNewsModal = newsPage.clickAddNewInfo();
-        editNewsModal.inputNewsTitle(createdTitle);
-        editNewsModal.inputNewsLinkTranslit(createdLink);
-        editNewsModal.inputNewsTextEditor(createdText);
-        editNewsModal.inputNewsCreationDate(new Date(System.currentTimeMillis()));
-
-        editNewsModal.clickUploadNewsPhoto(imagePath);
-        editNewsModal.waitUntilPhotoIsUploaded();
-
-        editNewsModal.saveNews();
     }
 
     @Test
-    @Issue("156")
-    public void testCreateNews() {
+    @Issue("169")
+    public void testCreateNewsEmptyFields() {
         NewsPageAdminPanel newsPage = new NewsPageAdminPanel(driver);
 
-        NewsPageGridComponent newsGrid = newsPage.getNewsPageGridComponent();
-        newsGrid.updateNewsRows(driver);
+        CreateEditNewsModal editNewsModal = newsPage.clickAddNewInfo();
 
-        NewsRowComponent createdNews = newsGrid.getRowById(0);
-        assertNotNull(createdNews, "Created news should exist.");
-        assertEquals(createdNews.getName().getText(), createdTitle, "The title is not correct.");
-        String expectedYear = String.valueOf(java.time.Year.now().getValue());
-        assertTrue(createdNews.getDateOfCreation().getText().contains(expectedYear), "Date was not correct.");
+        editNewsModal.inputNewsLinkTranslit(createdLink);
+        editNewsModal.inputNewsTextEditor(createdText);
+        editNewsModal.inputNewsCreationDate(createdDate);
+        editNewsModal.clickUploadNewsPhoto(imagePath);
+        assertFalse(editNewsModal.isSaveButtonEnabled());
 
-        String uploadedImageUrl = createdNews.getUploadedImageUrl(); 
+        editNewsModal.inputNewsTitle(createdTitle);
+        editNewsModal.inputNewsTextEditor(createdText);
+        editNewsModal.inputNewsCreationDate(createdDate);
+        editNewsModal.clickUploadNewsPhoto(imagePath);
+        assertFalse(editNewsModal.isSaveButtonEnabled());
 
-        assertTrue(uploadedImageUrl.contains(imagePath), "The uploaded image does not match the provided image.");
+        editNewsModal.inputNewsTitle(createdTitle);
+        editNewsModal.inputNewsLinkTranslit(createdLink);
+        editNewsModal.inputNewsCreationDate(createdDate);
+        editNewsModal.clickUploadNewsPhoto(imagePath);
+        assertFalse(editNewsModal.isSaveButtonEnabled());
+
+        editNewsModal.inputNewsTitle(createdTitle);
+        editNewsModal.inputNewsLinkTranslit(createdLink);
+        editNewsModal.inputNewsTextEditor(createdText);
+        editNewsModal.inputNewsCreationDate(createdDate);
+        assertFalse(editNewsModal.isSaveButtonEnabled());
+
+        editNewsModal.inputNewsTitle(createdTitle);
+        editNewsModal.inputNewsLinkTranslit(createdLink);
+        editNewsModal.inputNewsTextEditor(createdText);
+        editNewsModal.clickUploadNewsPhoto(imagePath);
+        assertFalse(editNewsModal.isSaveButtonEnabled());
     }
 
     @AfterMethod
